pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your SynthesisLab"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.12",
    "neo4j>=5.23.0",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "lxml>=5.3.0",
]
